requirements.txt
Django==1.10.5
djangorestframework==3.5.3
psycopg2==2.6.2
python-decouple==3.0
social-auth-app-django==1.0.0
grequests==0.3.0